Extract conversation window state
